+++ src/js/map/TwinTubeLayer.js
@@ -21,7 +21,7 @@ var _                  = require('lodash');
 /** 
 * Renders a topology object as d3 lines and circles into the provided svg element
 ```
-singleTubeLayer = SingleTubeLayer({
+twinTubeLayer = TwinTubeLayer({
         name: layer.name(),
         svg:  svg.append("g")
     })
@@ -40,15 +40,15 @@ singleTubeLayer = SingleTubeLayer({
         d3.select(d.event.target).style("cursor", "default");
     });
 ```
-* @class SingleTubeLayer
+* @class TwinTubeLayer
 * @extends BaseLayer
-* @constructor SingleTubeLayer
+* @constructor TwinTubeLayer
 * @static 
 * @property {Object} params - The parameter object for the Layer
 * @property {d3.selection(svg):Required} params.svg - A d3 selection of the svg element to render the editor layer into 
 */
 
-var SingleTubeLayer = function(params){
+var TwinTubeLayer = function(params){
     params.lineColor = params.lineColor || '#ddd';
     params.lineOpacity = params.lineOpacity || 1;
     params.endpointOpacity = params.endpointOpacity || 1;
@@ -66,13 +66,15 @@ var SingleTubeLayer = function(params){
     var svg = params.svg;
     svg.attr('id', layer.layerId());
     svg.classed(params.map.containerId(), true);
-    svg.classed('single-tube', true);
+    svg.classed('twin-tube', true);
+
+    //get leaflet map
+    var lmap = params.lmap;
 
     //define how to toggle the layer
     layer.onToggle(function(active){
         svg.transition().style('opacity', (active) ? 1 : 0);
     });
-
 
     var line = d3.svg.line()
     .interpolate("bundle")
@@ -81,6 +83,15 @@ var SingleTubeLayer = function(params){
         })
     .y(function(d) {
             return layer.latLngToXy([d.lat, d.lon])[1];
+        });
+    
+    var lineXY = d3.svg.line()
+    .interpolate("bundle")
+    .x(function(d) {
+            return d.x;
+        })
+    .y(function(d) {
+            return d.y;
         });
     
     //set default event handlers for links if none were passed in
@@ -185,7 +196,7 @@ var SingleTubeLayer = function(params){
     
         //make the scale of the arrow a function of the line width
         var arrow_scale;
-        if(layer.lineWidth() <=3){
+        if(layer.lineWidth() <=2){
             arrow_scale = 0;
         }else if(layer.lineWidth() <= 4){
             arrow_scale = 2;
@@ -195,16 +206,12 @@ var SingleTubeLayer = function(params){
             arrow_scale = 4;
         }
 
-        var path = d3.select(this.parentNode).select("path").filter(".adjacencyShadow")[0][0];
+        var path = d3.select(this.parentNode).select("path").filter(".adjacencyShadow1")[0][0];
         var p  = path.getPointAtLength(path.getTotalLength()/2);
-
+        
         //look after the midpoint in the path (away from A) to get our angle
         var p2 = path.getPointAtLength((path.getTotalLength()/2)+12);
-        //if Z is passing more traffic, look before the midpoint to get the angle away from Z
-        if(d.arrow == ARROW.ZA){
-            p2 = path.getPointAtLength((path.getTotalLength()/2)-12);
-        }
-        
+               
         var angle = getAngle(p,p2);
 
         //this works b/c it we know the ratio of the height/width of the arrow is 2 to 1
@@ -214,6 +221,61 @@ var SingleTubeLayer = function(params){
         p.x -= 2 * arrow_scale;
 
         return "translate(" + p.x+','+p.y+ ") scale("+arrow_scale+") rotate("+angle+",2,1)"
+     }
+
+    function arrowReverse(d){
+        if(d.arrow === undefined || d.arrow === ARROW.NONE){
+            return;
+        }
+
+        // arrow scaled to line width
+        var arrow_scale;
+        if(layer.lineWidth() <=2){
+            arrow_scale = 0;
+        }else if(layer.lineWidth() <=4){
+            arrow_scale = 2;
+        }else if(layer.lineWidth() <=6){
+            arrow_scale = 3;
+        }else{
+            arrow_scale = 4;
+        }
+    
+        var path = d3.select(this.parentNode).select(".adjacencyShadow2")[0][0];
+        var p = path.getPointAtLength(path.getTotalLength()/2);
+
+        p2 = path.getPointAtLength((path.getTotalLength()/2)+12);
+
+        var angle = getAngle(p,p2);
+
+        p.y -= 1 * arrow_scale;
+        p.x -= 2 * arrow_scale;
+
+        return "translate(" +p.x+","+p.y+") scale("+arrow_scale+") rotate("+angle+",2,1)";
+
+    }
+
+    // Helper function to convert lat/lngs to x,y coordinates
+    function _geoToXY(path){
+
+        let newPath = [];
+        for(let i = 0; i<path.length; i++){
+            let point = lmap.latLngToLayerPoint(L.latLng(path[i]["lat"], path[i]["lon"]));
+            newPath.push(point); 
+        }
+        return newPath;
+    }
+    
+    // Helper function to shift path #1
+    function _pathShifter(newPath, theta, pathWidth){
+        let shiftedPath = [];
+        let wsin = pathWidth * Math.sin(theta);
+        let wcos = pathWidth * Math.cos(theta);
+        for(let i = 0; i< newPath.length; i++){
+            let newX = newPath[i].x - wsin;
+            let newY = newPath[i].y - wcos;
+            shiftedPath.push({"x": newX, "y": newY});
+        }
+        return shiftedPath;
     }
 
     //define how to update the layers components
@@ -222,6 +284,9 @@ var SingleTubeLayer = function(params){
             console.warn('No topology set, skipping update for '+layer.name());
             return;
         }
+
+        var pathWidth = layer.lineWidth();
+
         //--- Render Links
         var links = layer.links()
             .data(layer.topology().data().links, function(d){
@@ -234,13 +299,20 @@ var SingleTubeLayer = function(params){
             .attr("id", function(d) { return d.linkId })
             .attr("class","adjacency");
 
-        //add a shadow path for new adjacencys 
-        linksEnter.append("path")
-        .attr("d",function(d){
-            return line(d.path);
-        })
-        //.classed('adjacencyShadow adjacencyHighlight', true)
-            .attr("class","adjacencyShadow")
+        //add a shadow path #1 for new adjacencys 
+        var adjacencyShadow1 = linksEnter.append("path")
+            .attr("d",function(d){ 
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(y1-yn, xn-x1);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth);         
+                return lineXY(shiftedPath);
+            })
+            .attr("class","adjacencyShadow1")
             .call(function(selection){
                 _.forEach(layer.onLinkEvent(), function(callback, evt){
                     selection.on(evt, function(d){
@@ -252,10 +324,46 @@ var SingleTubeLayer = function(params){
                 });
             });
 	
-        //add a highlight path for new links
-        linksEnter.append("path")
-            .attr("d",function(d){return line(d.path)})
-            .attr("class","adjacencyHighlight")
+        //add a highlight path #1 for new links
+        var adjacencyHighlight1 = linksEnter.append("path")
+            .attr("d",function(d){
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(y1-yn, xn-x1);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth); 
+                return lineXY(shiftedPath);
+            })
+            .attr("class","adjacencyHighlight1")
+            .call(function(selection){
+                _.forEach(layer.onLinkEvent(), function(callback, evt){
+                    selection.on(evt, function(d){
+                        callback({
+                            event: d3.event,
+                            data:  d
+                        });
+                    });
+                });
+            });
+
+        //add a shadow path #2 for new adjacencys 
+        var adjacencyShadow2 = linksEnter.append("path")
+            .attr("d",function(d){ 
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(yn-y1, x1-xn);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth);
+                _.reverse(shiftedPath);    
+                return lineXY(shiftedPath);
+            })
+            .attr("class","adjacencyShadow2")
             .call(function(selection){
                 _.forEach(layer.onLinkEvent(), function(callback, evt){
                     selection.on(evt, function(d){
@@ -267,52 +375,196 @@ var SingleTubeLayer = function(params){
                 });
             });
 	
+        //add a highlight path #2 for new links
+        var adjacencyHighlight2 = linksEnter.append("path")
+            .attr("d",function(d){
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(yn-y1, x1-xn);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth); 
+                _.reverse(shiftedPath);
+                return lineXY(shiftedPath);
+            })
+            .attr("class","adjacencyHighlight2")
+            .call(function(selection){
+                _.forEach(layer.onLinkEvent(), function(callback, evt){
+                    selection.on(evt, function(d){
+                        callback({
+                            event: d3.event,
+                            data:  d
+                        });
+                    });
+                });
+            });
+
         //add directional indicator by appending yet another path to the g
         linksEnter.append("path")
-            .attr("class","arrow")
+            .attr("class","arrow1")
             .attr("d","M.5,1 L0,2 L3,1 L0,0 Z")
             .attr("transform", arrowTranslate);
-	
+        
+        linksEnter.append("path")
+            .attr("class","arrow2")
+            .attr("d","M.5,1 L0,2 L3,1 L0,0 Z")
+            .attr("transform", arrowReverse);    
+        
         //--- UPDATE -- update the paths of any existing links
-	
-        //update shadow path
-        links.select(".adjacencyShadow")
-            .attr("d",function(d){return line(d.path)})
+
+        //update shadow path #1
+        links.select(".adjacencyShadow1")
+            .attr("d",function(d){
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(y1-yn, xn-x1);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth); 
+                return lineXY(shiftedPath);
+            })
             .style('stroke-width', function(d){
-		var strokeWidth = d.lineColor === undefined ? (layer.lineWidth()-1) : (layer.lineWidth());
+		        var strokeWidth = d.lineColor === undefined ? (layer.lineWidth()-1) : (layer.lineWidth());
                 if (d.selected) {
+                    pathWidth = strokeWidth * 2;
                     return (strokeWidth * 2)+'px';
                 } else {
+                    pathWidth = strokeWidth;
                     return strokeWidth+'px';
                 }
             })
             .style('opacity',function(d){
+                if(d.azLineOpacity){
+                    return d.azLineOpacity;
+                }
                 return d.lineOpacity === undefined ? layer.lineOpacity() : d.lineOpacity;
             })
             .style('stroke', function(d){
+                if(d.azLineColor){
+                    return d3.rgb(d.azLineColor).darker();
+                }
                 return d.lineColor === undefined ? d3.rgb(layer.lineColor()).darker() : d3.rgb(d.lineColor).darker();
             });
-        //update highlight path
-        links.select(".adjacencyHighlight")
-            .attr("d",function(d){return line(d.path)})
+        //update highlight path #1
+        links.select(".adjacencyHighlight1")
+            .attr("d",function(d){
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(y1-yn, xn-x1);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth); 
+                return lineXY(shiftedPath);
+            })
             .style('stroke-width', function(d){
                 var strokeWidth = d.lineColor === undefined ? (layer.lineWidth()-2) : (layer.lineWidth()-1);
                 if (d.selected) {
+                    pathWidth = strokeWidth * 2;
                     return (strokeWidth * 2)+'px';
                 } else {
+                    pathWidth = strokeWidth;
                     return strokeWidth+'px';
                 }
             })
             .style('opacity', function(d){
+                if(d.azLineOpacity){
+                    return d.azLineOpacity;
+                }
                 return d.lineOpacity === undefined ? layer.lineOpacity() : d.lineOpacity;
             })
             .style('stroke', function(d){
+                if(d.azLineColor){
+                    return d3.rgb(d.azLineColor);
+                }
                 return d.lineColor === undefined ? d3.rgb(layer.lineColor()) : d3.rgb(d.lineColor);
             });
-        
+
+
+        //update shadow path #2
+        links.select(".adjacencyShadow2")
+            .attr("d",function(d){
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(yn-y1, x1-xn);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth);
+                _.reverse(shiftedPath); 
+                return lineXY(shiftedPath);
+            })
+            .style('stroke-width', function(d){
+		        var strokeWidth = d.lineColor === undefined ? (layer.lineWidth()-1) : (layer.lineWidth());
+                if (d.selected) {
+                    pathWidth = strokeWidth * 2;
+                    return (strokeWidth * 2)+'px';
+                } else {
+                    pathWidth = strokeWidth;
+                    return strokeWidth+'px';
+                }
+            })
+            .style('opacity',function(d){
+                if(d.zaLineOpacity){
+                    return d.zaLineOpacity;
+                }
+                return d.lineOpacity === undefined ? layer.lineOpacity() : d.lineOpacity;
+            })
+            .style('stroke', function(d){
+                if(d.zaLineColor){
+                    return d3.rgb(d.zaLineColor).darker();
+                }
+                return d.lineColor === undefined ? d3.rgb(layer.lineColor()).darker() : d3.rgb(d.lineColor).darker();
+            });
+        //update highlight path #2
+        links.select(".adjacencyHighlight2")
+            .attr("d",function(d){
+                let newPath = _geoToXY(d.path);
+                let end_idx = newPath.length-1;
+                let x1 = newPath[0].x;
+                let y1 = newPath[0].y;
+                let xn = newPath[end_idx].x;
+                let yn = newPath[end_idx].y;
+                let theta = Math.atan2(yn-y1, x1-xn);
+                let shiftedPath = _pathShifter(newPath, theta, pathWidth);
+                _.reverse(shiftedPath);
+                return lineXY(shiftedPath);
+            })
+            .style('stroke-width', function(d){
+                var strokeWidth = d.lineColor === undefined ? (layer.lineWidth()-2) : (layer.lineWidth()-1);
+                if (d.selected) {
+                    pathWidth = strokeWidth * 2;
+                    return (strokeWidth * 2)+'px';
+                } else {
+                    pathWidth = strokeWidth;
+                    return strokeWidth+'px';
+                }
+            })
+            .style('opacity', function(d){
+                if(d.zaLineOpacity){
+                    return d.zaLineOpacity;
+                }
+                return d.lineOpacity === undefined ? layer.lineOpacity() : d.lineOpacity;
+            })
+            .style('stroke', function(d){
+                if(d.zaLineColor){
+                    return d3.rgb(d.zaLineColor);
+                }
+                return d.lineColor === undefined ? d3.rgb(layer.lineColor()) : d3.rgb(d.lineColor);
+            });
+
         //update arrow path
-        links.select('.arrow')
+        links.select('.arrow1')
             .attr("transform",arrowTranslate);
+
+        links.select('.arrow2')
+            .attr("transform", arrowReverse);
 	
         //--- EXIT -- remove any links we no longer need
         links.exit().remove();
@@ -372,19 +624,18 @@ var SingleTubeLayer = function(params){
             });
 
 
-
         //--- EXIT -- remove any endpoint we no longer need
         endpoints.exit().remove();
 	
-	if(!this.isInitDone()){
-	    console.log("Init Complete!");
-	    this.isInitDone(true);
-	    this.initComplete();
-	}
+	    if(!this.isInitDone()){
+	        console.log("Init Complete!");
+	        this.isInitDone(true);
+	        this.initComplete();
+	    }
 
         return true;
     });
 
     return layer;
 };
-module.exports = SingleTubeLayer;
+module.exports = TwinTubeLayer;
